Corrected schemas for user

+++ api/blueprints/admin/routes.py
@@ -12,6 +12,7 @@ from schemas.course.add_course_schema import AddCourseSchema
 from schemas.course.add_current_course_schema import AddCurrentCourseSchema
 from schemas.course.update_course_schema import UpdateCourseSchema
 from schemas.course.update_current_course import UpdateCurrentCourseSchema
+from schemas.user.update_user_schema import UpdateUserSchema
 from schemas.user.add_user_schema import AddUserSchema
 from utils.file_utils import allowed_file, allowed_file_size
 from utils.get_file_url import get_file_url
@@ -20,32 +21,32 @@ from utils.save_uploaded_file import save_uploaded_file
 
 admin_bp = Blueprint('admin', __name__)
 
-
+# Update current course schema
 update_current_course_schema = UpdateCurrentCourseSchema()
+
 # UPDATE
 @admin_bp.route("/update-current-course/<int:id>", methods=['PUT'])
 @jwt_required()
 @role_required(["admin"])
 def update_current_course(id):
     try:
+        con, cursor= None, None
         con, cursor = get_db_connection()
 
-        # Uzmi JSON podatke
+        #JSON podaci
         data = request.json 
 
-        # Uzmi trenutne vrednosti iz baze
         cursor.execute("SELECT * FROM current_courses WHERE id=%s", (id,))
         current_course = cursor.fetchone()
         if not current_course:
             return jsonify({"message": "Current course not found"}), 404
 
-        # Validacija opcionih polja
         try:
             validated_data = update_current_course_schema.load(data)
         except ValidationError as err:
             return jsonify({"errors": err.messages}), 400
 
-        # Kombinuj stare i nove vrednosti
+        # Combine old and new values
         professor = validated_data.get('user_id', current_course['user_id'])
         price = validated_data.get('price', current_course['price'])
         start_at = validated_data.get('start_at', current_course['start_at'])
@@ -56,7 +57,7 @@ def update_current_course(id):
         lessons = validated_data.get('lessons', current_course['lessons'])
         course_id =  current_course['course_id']
 
-        #  provera da li profesor i kurs postoje, ako su poslati
+        # checking if the professor and the course exist
         if 'user_id' in validated_data:
             cursor.execute("SELECT id FROM user WHERE id=%s AND rola='professor'", (professor,))
             if not cursor.fetchone():
@@ -67,7 +68,7 @@ def update_current_course(id):
             if not cursor.fetchone():
                 return jsonify({"message": f"Course with id {course_id} does not exist"}), 400
 
-        # Provera preklapanja termina ako je start_at, end_at ili profesor poslato
+        # Check for appointment overlap if start_at, end_at or professor is sent
         if any(k in validated_data for k in ['start_at', 'end_at', 'user_id', 'course_id']):
             cursor.execute("""
                 SELECT id FROM current_courses
@@ -100,10 +101,7 @@ def update_current_course(id):
         if con:
             con.close()
 
-
-img_base_url_user = 'http://127.0.0.1:5000/uploads/user/'
-UPLOAD_FOLDER_USER = 'uploads/user'
-
+# update_user_schema = UpdateUserSchema()
 
 @admin_bp.route("/update-user/<int:id>", methods=['PUT'])
 @jwt_required()
@@ -112,33 +110,69 @@ def update_user_info(id):
     try:
         con, cursor = get_db_connection()
 
+        upload_folder_user = current_app.config["UPLOAD_FOLDER_USER"]
+        allowed_extensions = current_app.config["ALLOWED_IMAGE_EXTENSIONS"]
+
         query = "SELECT  * FROM user WHERE id=%s"
         cursor.execute(query, (id,))
         user = cursor.fetchone()
 
+        if not user:
+            return jsonify({"error": "User not found"}), 404
+        print(type(user))
+        schema = UpdateUserSchema()
+        schema.context["current_user"] = user
+
         file = request.files.get('file')
-
-        if file:
-            file_url = str(img_base_url_user+file.filename)
-            file.save(os.path.join(UPLOAD_FOLDER_USER, file.filename))
-
-        else:
-            file_url = user['user_image_url']
-
         data = request.form.to_dict()
 
-        firstName = data.get('first_name', user['first_name'])
-        lastName = data.get('last_name', user['last_name'])
-        phoneNumber = data.get('phone_number', user['phone_number'])
-        email = data.get('email', user['email'])
-        biography = data.get('biography', user['biography'])
-        rola = data.get('rola', user['rola'])
-        password = data.get('password')
+        try:
+            validated_data = schema.load(
+            data,
+            unknown="exclude"  
+            )
+        except ValidationError as err:
+            return jsonify({"errors": err.messages}), 400
+        
+        filename = user['user_image_url']
+
+        if file:
+            old_filename =  user['user_image_url']
+            # Check extension
+            if not allowed_file(file.filename):
+                allowed_extensions = current_app.config.get("ALLOWED_IMAGE_EXTENSIONS", set())
+                return jsonify({
+                "message": f"Invalid image format. Allowed: {', '.join(allowed_extensions)}"
+                }), 400
+
+            # Check file size (max 2MB)
+            if not allowed_file_size(file, max_size_mb=2):
+                return jsonify({"message": "File is too large. Max size is 2MB."}), 400
+
+            # Save new image
+            filename = save_uploaded_file(file,upload_folder_user)
+
+            # Remove old image
+            if old_filename and old_filename != "default_user.png":
+                old_path = os.path.join(upload_folder_user, old_filename)
+                if os.path.exists(old_path):
+                    os.remove(old_path)
+            
+       
+        password = validated_data.get("password")
 
         if password:
-            hashed_password = generate_password_hash(password)
+            password_hashed = generate_password_hash(password)
         else:
-            hashed_password = user['password_hash']
+            password_hashed = user['password_hash']
+    
+        first_name = validated_data.get('first_name', user['first_name'])
+        last_name = validated_data.get('last_name', user['last_name'])
+        phone_number = validated_data.get('phone_number', user['phone_number'])
+        email = validated_data.get('email', user['email'])
+        biography = validated_data.get('biography', user['biography'])
+        rola = validated_data.get('rola', user['rola'])
+        password = password_hashed
 
         query = """
         UPDATE user
@@ -146,8 +180,8 @@ def update_user_info(id):
         WHERE id = %s
         """
 
-        values = (firstName, lastName, email, phoneNumber,
-                  biography, file_url, hashed_password, rola, id)
+        values = (first_name, last_name, email, phone_number,
+                  biography, filename, password_hashed, rola, id)
 
         cursor.execute(query, values)
         con.commit()
@@ -156,6 +190,12 @@ def update_user_info(id):
     except Exception as e:
         print(f"An unexpected error occurred: {e}")
         return jsonify({"error": "Internal server error"}), 500
+    
+    finally:
+        if cursor:
+            cursor.close()
+        if con:
+            con.close()
 
 # Course update Schema for Validation
 update_course_schema = UpdateCourseSchema()
@@ -181,7 +221,7 @@ def update_course(id):
         try:
             validated_data = update_course_schema.load(
             data,
-            unknown="exclude"  # ili unknown=EXCLUDE
+            unknown="exclude"  
             )
         except ValidationError as err:
             return jsonify({"errors": err.messages}), 400
@@ -199,7 +239,7 @@ def update_course(id):
             if not allowed_file_size(file, max_size_mb=2):
                 return jsonify({"message": "File is too large. Max size is 2MB."}), 400
 
-       
+            # Save new image
             filename = save_uploaded_file(file,upload_folder_course)
 
             # Remove old image
@@ -262,6 +302,15 @@ def get_all_users():
         """
         cursor.execute(query)
         data = cursor.fetchall()
+
+        # generate img url
+        for user in data:
+            user["user_image_url"] = get_file_url(
+            user.get("user_image_url"),
+            folder_type="user",
+            default="default_user.png"
+        )
+            
         return jsonify(data)
 
     except Exception as e:
@@ -298,6 +347,11 @@ def get_user(id):
         if not data:
             return jsonify({"message": "User not found"}), 404
         
+        data["user_image_url"] = get_file_url(
+            data.get("user_image_url"),
+            folder_type="user",
+            default="default_user.png")
+        
         return jsonify(data)
 
     except Exception as e:
@@ -317,10 +371,6 @@ def get_all_current_courses(id):
     try:
         con, cursor = get_db_connection()
         
-
-        # level = request.args.get('level', '%')
-        # if (level == ""):
-        #     level = "%"
         level = request.args.get('level') or '%'
 
         query = """
@@ -339,6 +389,7 @@ def get_all_current_courses(id):
         data = cursor.fetchall()
         print(data)
 
+        # generate img url
         for course in data:
             course["course_image_url"] = get_file_url(
             course.get("course_image_url"),
@@ -363,7 +414,7 @@ def get_all_current_courses(id):
 @jwt_required()
 @role_required(["admin"])
 def get_all_courses():
-    con, cursor = None, None  # Inicijalizujemo konekciju i cursor
+    con, cursor = None, None  
     try:
         con, cursor = get_db_connection()
 
@@ -449,7 +500,7 @@ def get_professors():
 
     except Exception as e:
         print(f"An unexpected error occurred: {e}")
-        return jsonify({"message": "An error occurred while fetching courses."}), 500
+        return jsonify({"message": "An error occurred while fetching professors."}), 500
     finally:
         if cursor:
             cursor.close()
@@ -466,7 +517,8 @@ add_user_schema = AddUserSchema()
 @role_required(["admin"])
 def add_user():
     try:
-
+        con = None
+        cursor = None
         con, cursor = get_db_connection()
 
         upload_folder_user = current_app.config["UPLOAD_FOLDER_USER"]
@@ -476,24 +528,26 @@ def add_user():
         file = request.files.get('file')
         data = request.form.to_dict()
 
-        # Validating text fields with Marshmallow
-        try:
-            data = add_user_schema.load(request.form)
-        except ValidationError as err:
-            return jsonify({"errors": err.messages}), 400
-        
-        # Provera da li email postoji
+        # Check if email exists
         cursor.execute("SELECT id FROM user WHERE email = %s", (data["email"],))
         if cursor.fetchone():
             return jsonify({"message": "User with this email already exists!"}), 409
 
-        password_hash = generate_password_hash(data['password'])
+        print(request.form.to_dict())
+
+        # Validating text fields with Marshmallow
+        try:
+            validated_data = add_user_schema.load(data)
+        except ValidationError as err:
+            return jsonify({"errors": err.messages}), 400
+        
+        password_hash = generate_password_hash(validated_data['password'])
 
         # Image is required only for professors
-        if data['rola'] == 'professor' and (not file or file.filename == ""):
+        if validated_data['rola'] == 'professor' and (not file or file.filename == ""):
             return jsonify({"message": "Image is required!"}), 400
 
-        # Ako fajl postoji → sačuvaj ga
+       # If the file exists save it
         if file and file.filename != "":
             # Check exte.
             if not allowed_file(file.filename):
@@ -502,17 +556,16 @@ def add_user():
                 "message": f"Invalid image format. Allowed: {', '.join(allowed_extensions)}"
                 }), 400
 
-        # Provera veličine fajla (max 2MB)
+        # Check file size (max 2MB)
             if not allowed_file_size(file, max_size_mb=2):
                 return jsonify({"message": "File is too large. Max size is 2MB."}), 400
         
-            file_url = save_uploaded_file(file, upload_folder_user,
-                              img_base_url_user)
+            filename = save_uploaded_file(file,upload_folder_user)
 
-        # Ako nema fajla i nije professor → stavlja default
+        # If there is no file and it is not a professor,it sets the default
         else:
             filename = current_app.config["DEFAULT_USER_IMAGE"]
-            file_url = f"{img_base_url_user.rstrip('/')}/{filename}"
+           
 
         query = """
         INSERT INTO user 
@@ -520,14 +573,14 @@ def add_user():
         VALUES (%s, %s, %s, %s, %s, %s, %s, %s)
         """
         values = (
-            data['first_name'],
-            data['last_name'],
-            data['email'],
-            data.get('phone_number'),
-            data.get('biography'),
+            validated_data['first_name'],
+            validated_data['last_name'],
+            validated_data['email'],
+            validated_data['phone_number'],
+            validated_data['biography'],
             password_hash,
-            data['rola'],
-            file_url
+            validated_data['rola'],
+            filename
         )
 
         cursor.execute(query, values)
@@ -538,6 +591,7 @@ def add_user():
     except Exception as e:
         print(f"An unexpected error occurred: {e}")
         return jsonify({"message": "Error!"})
+    
     finally:
         if cursor:
             cursor.close()
@@ -694,6 +748,21 @@ def delete_course(id):
 
         con, cursor = get_db_connection()
 
+        upload_folder_course = current_app.config["UPLOAD_FOLDER_COURSE"]
+
+        cursor.execute("SELECT course_image_url FROM course WHERE id = %s", (id,))
+        course = cursor.fetchone()
+
+        if not course:
+            return jsonify({"message": "Course not found."}), 404
+        
+        filename = course['course_image_url']
+
+        if filename and filename != "default_user.png":
+                old_path = os.path.join(upload_folder_course, filename)
+                if os.path.exists(old_path):
+                    os.remove(old_path)
+                    
         query = """
         DELETE FROM course WHERE id = %s;
         """
@@ -751,6 +820,7 @@ def delete_current_course(id):
             con.close()
 
 
+
 @admin_bp.route("/delete-user/<int:id>", methods=['DELETE'])
 @jwt_required()
 @role_required(["admin"])
@@ -758,6 +828,21 @@ def delete_user(id):
     try:
 
         con, cursor = get_db_connection()
+
+        upload_folder_user = current_app.config["UPLOAD_FOLDER_USER"]
+
+        cursor.execute("SELECT user_image_url FROM user WHERE id = %s", (id,))
+        user = cursor.fetchone()
+
+        if not user:
+            return jsonify({"message": "User not found."}), 404
+        
+        filename = user['user_image_url']
+
+        if filename and filename != "default_user.png":
+                old_path = os.path.join(upload_folder_user, filename)
+                if os.path.exists(old_path):
+                    os.remove(old_path)
 
         query = """
         DELETE FROM user WHERE id = %s;
@@ -783,4 +868,3 @@ def delete_user(id):
         if con:
             con.close()
 
-#to do on delete delete images
